[PATCH] PATCH [6/15]  qla2xxx: LoopID downcast fix
  Fix problem where the driver would incorrectly down-cast the target
  loop_id while retrieving link statistics.

 drivers/scsi/qla2xxx/qla_gbl.h |    2 +-
 drivers/scsi/qla2xxx/qla_mbx.c |    2 +-
 2 files changed, 2 insertions(+), 2 deletions(-)

+++ drivers/scsi/qla2xxx/qla_gbl.h
@@ -198,7 +198,7 @@ extern int
 qla2x00_get_port_name(scsi_qla_host_t *, uint16_t, uint8_t *, uint8_t);
 
 extern uint8_t
-qla2x00_get_link_status(scsi_qla_host_t *, uint8_t, link_stat_t *, uint16_t *);
+qla2x00_get_link_status(scsi_qla_host_t *, uint16_t, link_stat_t *, uint16_t *);
 
 extern int
 qla2x00_lip_reset(scsi_qla_host_t *);
